refactor: refactor my-requests
